+++ gui/PopUpWidget.cpp
@@ -307,7 +307,7 @@ void PopUpWidget::clearEntries() {
 void PopUpWidget::setSelected(int item) {
 	// FIXME
 	if (item != _selectedItem) {
-		if (item >= 0 && item < _entries.size()) {
+		if (item >= 0 && item < (int)_entries.size()) {
 			_selectedItem = item;
 		} else {
 			_selectedItem = -1;
